+++ 2025/day03/functions.h
@@ -3,7 +3,7 @@
 
 #include <string>
 
-size_t max(const std::string &line, const int num_bat);
+long long max(const std::string &line, const int num_bat);
 
 
 #endif // FUNCTIONS_H
